requirements.txt
apsw==3.44.0.0
certifi==2023.7.22
charset-normalizer==3.3.2
idna==3.4
pycryptodome==3.19.0
PyYAML==6.0.1
requests==2.31.0
urllib3==2.1.0